+++ lib/arachni/rpc/client/instance/framework.rb
@@ -26,9 +26,9 @@ class Framework < Proxy
     %w(list_reporters list_plugins).each do |m|
         translate m do |data|
             data.map do |c|
-                c = c.symbolize_keys
+                c = c.my_symbolize_keys
                 c[:options] = c[:options].map do |o|
-                    o = o.symbolize_keys
+                    o = o.my_symbolize_keys
                     o[:name] = o[:name].to_sym
                     o[:type] = o[:type].to_sym
                     o
@@ -39,18 +39,18 @@ class Framework < Proxy
     end
 
     translate :list_platforms do |platforms|
-        platforms.inject({}) { |h, (k, v)| h[k] = v.symbolize_keys; h }
+        platforms.inject({}) { |h, (k, v)| h[k] = v.my_symbolize_keys; h }
     end
 
     translate :statistics do |stats|
-        stats.symbolize_keys
+        stats.my_symbolize_keys
     end
 
     translate :progress do |data, options = {}|
         sitemap = data.delete('sitemap')
         issues  = data.delete('issues')
 
-        data = data.symbolize_keys
+        data = data.my_symbolize_keys
         data[:status] = data[:status].to_sym
 
         if issues
@@ -62,7 +62,7 @@ class Framework < Proxy
         end
 
         if data[:instances]
-            data[:instances] = data[:instances].map(&:symbolize_keys)
+            data[:instances] = data[:instances].map(&:my_symbolize_keys)
         end
 
         if sitemap