+++ spec/arachni/ruby/hash_spec.rb
@@ -23,14 +23,14 @@ describe Hash do
         }
     end
 
-    describe '#stringify_keys' do
+    describe '#my_stringify_keys' do
         it 'recursively converts keys to strings' do
-            with_symbols.stringify_keys.should == with_strings
+            with_symbols.my_stringify_keys.should == with_strings
         end
 
         context 'when the recursive is set to false' do
             it 'only converts the keys at depth 1' do
-                with_symbols.stringify_keys( false ).should == {
+                with_symbols.my_stringify_keys( false ).should == {
                     'stuff' => 'blah',
                     'more'  => {
                         stuff: {
@@ -42,14 +42,14 @@ describe Hash do
         end
     end
 
-    describe '#symbolize_keys' do
+    describe '#my_symbolize_keys' do
         it 'recursively converts keys to symbols' do
-            with_strings.symbolize_keys.should == with_symbols
+            with_strings.my_symbolize_keys.should == with_symbols
         end
 
         context 'when the recursive is set to false' do
             it 'only converts the keys at depth 1' do
-                with_strings.symbolize_keys( false ).should == {
+                with_strings.my_symbolize_keys( false ).should == {
                     stuff: 'blah',
                     more:  {
                         'stuff' => {
@@ -75,14 +75,14 @@ describe Hash do
         end
     end
 
-    describe '#stringify' do
+    describe '#my_stringify' do
         it 'returns a Hash with keys and values recursively converted to strings' do
             {
                 test:         'blah',
                 another_hash: {
                     stuff: 'test'
                 }
-            }.stringify.should == {
+            }.my_stringify.should == {
                 'test'         => 'blah',
                 'another_hash' => {
                     'stuff' => 'test'